Clear out mail folder if already exists
I'm finding a lot of test runs of this job are filling the mail folder with files and subsequent runs take an extremely long time to try to parse through them all. Better to be proactive and clear the folder in the event a previous command failed/was stopped and left orphan files.