New Check - Agent Mail Profile
Checks where checking that Mail XPs and a Mail Profile existed but not the the SQL Agent had been configured to use said mail profile.

Use Get-DbaAgentServer to pull out the configured DatabaseMailProfile and check it against the required database mail profile.

Systems not enabled will output empty so will fail against $null or a configured value.

This check is needed as without it you cant send alerts from the agent so the checks for email enabled alerts on Sev16-25 etc need this to be on.